DEVSOL-903: Purchased Plans Page: Display plans grouped by status
The purchased plans page now uses a formatted table that is sortable, and by default is sorted by status.  Refactored the dates attributes of the RatePlan and DeveloperRatePlan objects to encapulate the business logic of getting dates, and all getters for dates now return DateTime objects instead of strings.

Conflicts:
	Apigee/Mint/DeveloperRatePlan.php